Revert "SI-6811 Misc. removals in util, testing, io, ..."
This partially reverts commit ea517bc610672c794091dad488316a6f1a6a0446.

The commit got reverted because it breaks Sbt that relies on
the old implementation of MurmurHash. The new implementation
got introduced in Scala 2.10 but sbt supports Scala 2.9 so
there's no way to migrate it to the new implementation hence
we have to keep the old one a while longer.

Review by @paulp