Revert "SI-6811 Misc. removals in util, testing, io, ..."
This partially reverts commit f2c784ee41b387b41be023338efe32f106f149b9.

The commit got reverted because it breaks Sbt that relies on
the old implementation of MurmurHash. The new implementation
got introduced in Scala 2.10 but sbt supports Scala 2.9 so
there's no way to migrate it to the new implementation hence
we have to keep the old one a while longer.

Review by @paulp