Pre-initializing Unpickler NamedTypes with their Denotations.
This fixes a weird problem where we read an external reference to a type parameter with NoPrefix as owner.
The problem arises when reading reflect.io.ZipFileArchive, where the reference is to the type parameter of AbstactFile.
It might be a known problem for Scala. For the moment there's nothing much we can do on the dotty side to deal with it. We either use .withDenot, and accept that this cannot be reloaded in future runs, or we use a HasFixedSym and risk a stale reference. It seems better to fail than to live with stale references.