SI-7265 javaSpecVersion, adjust isJava... tests
Make this and related properties public, because they
are useful.

This change ought to have been committed at 2.11 and then
backported with restrictions, rather than vice-versa.

Note that they are defined in the order, version, vendor
and name, which is the order from the underlying javadoc.

It would be a neat feature of the PR validator, as previously
imagined, to run a "pending" test and then, on success and
merge, to move it automatically to the canonical suite.