Eliminating getter generation in namer/typer.
The idea is that getters shoul dbe generated later (probably during UnCurry). This is much better for handling modules,
and also obviates the need to have LocalNames. Getters can be generated as soon in the compilation pipeline as methoda always get a parameter list. That way, we can distinguihs getters from fields without playing any tricks with names.