Changed type comparers to compare symbols of standard types instead of types themselves
Types themselves are not unique, for instance, scala.Any would be

    TypeRef(ThisType(scala), Any)

or

    TypeRef(TermRef(ThisType(_root_), scala), Any)

Even if we pick only one internally we still risk unpickling both forms when reading scala 2 - compiled files.
The only way to avoid this would be to change the ThisType constructor to always return a TypeRef for a module.