Following type aliases when pattern matching.
Faced with a pattern like

  Apply(x, xs)

we first look for an Apply object which is an extractor. If this fails, we now
interprete Apply as a type. If it is a type alias which points to a class type that
has a companion module, we then try that companion module as an extractor.

Scala 2.x does ot that way, and it's used widely within dotty itself. Think tpd.Apply as the found object,
Trees.Apply as the extractor.

Also, added a fix to normalization which made normalization go deep into a method type.