Change to computing arguments in a pattern match.
Previously, if the result of an unapply arg type is a type with isDefined and get methods, we proceed as follows:

  if the get result type is a product: take the produce argument types
  otherwise take the get result type itself

The problem is if the get result type is an empty product. We solve this by demanding that the get result type is a ProductN type, not just a subclass of product.