Adding () insertion to normalization.
We need to take account of the possibility of inserting a () argument list in normalization. Otherwise, a type with a

    def toString(): String

member would not count as a valid solution for ?{toString: String}. This would then lead to an implicit insertion, with a nice explosion of inference search because of course every implicit result has some sort
of toString method.

The problem is solved by dereferencing nullary method types if the corresponding function type is not
compatible with the prototype.