Fixing desugarings of classes.
1) get Method

The idea is that if a case class has exactly one case field, then that field is returned as result of the get method. If it has more than one, it is a subtype of Product and the fields are returned by the _i methods.

2) unapply: Boolean

If a class has no case fields the unapply method returns true, rather than the this.

3) Self names

A self clause

    { self => ...

if now desugared to

    { self: C =>

where C is the type of the class.