Taking a fully-defined result type into account when forming a closure.
Given

     (params) => body

and an expected type

     Ts => R

with R fully defined, we use R as the result type of the closure def into which the function is translated.

Note: It would be nice generalize this further to result types which are not fully defined, yet this would complicate things by entangling typing and desugaring.