Fixing a type problem where code does not compile under dotty.
... and I believe should not compile under Scala2x either. The problem is in line 361 of TreeInfo.scala

    methPart(tree) match { ...

Here,

    tree: tpd.Tree
    methPart: (tree: this.Tree): Tree

So we need to show that

    tpd.Tree <: this.Tree

LHS expands to

    ast.Tree[Type]

RHS expands to

    ast.Tree[T]   where T >: Untyped is TreeInfo's type parameter

Since Tree is contravariant, we need to etablish

    T <: Type

but I see nothing that could prove this. The Dotty typechecker detected the problem, yet Scala2x's didn't. Need to follow up on why not. For now, adding the necessary constraint to the codebase.