Relax type checking rule for pattern matching.
Previously, if S is the selector type and P is the type of the unapply argument, we demanded either

     S <: P

or

     P <: S

We now also admit P <: S' where S' can be obtained from S by dropping refinements. This lets some patterns in RefinedPrinters typecheck which did not typecheck before. It seems we can afford to be a little bit more liberal here, and thereby approximate the weaker "isPopulated" criterion of Scala2x.