Fixing problems around isAbsent
A Symbol is absent if (1) it equals NoSymbol, or (2) it's type is NoType or (3) it is a module with an module class which is absent.

Note that isAbsent does not force symbols. The info has to be completed explicitly to check.

Criterion (3) is missing, we now take it into account.

Also, made ClassDenotation ops more robost in tge face of possibilty of being absent.

Finally, refined logic in Typer, to complete a symbol before checking absent.