Discard inaccessible denotations when merging.
When froming the &=merge of two denotations, we now discard denotations which have inaccessible symbols.

The reason for this change is that, without it, the following code gives the warning "cannot merge Int and Int() as members of one type, keeping only Int()".

    val m = new java.util.HashMap
    m.size

Indeed java.util.HashMap has a filed and method which are both named size. The field is package private, so is inaccessible from outside java.util, however.