Move dotty compiler code to compiler.internal
This commit moves all code that was in dotty.tools into
dotty.compiler.internal. The idea of a "tools" package seems to be a
holdover from "com.sun.tools.javac". This naming style is not in line
with how packages are normally named in the standard Scala distribution,
hence the change. Note that I've done my best to update tests to make
them compile/run, but some tests that did not compile before still don't
compile with this commit. These tests are not in the set run by
"sbt test", so I believe them to be stale (in fact, they should likely
be removed in a future commit).