Fix for type applications.
Test case in TreeTransformers.scala. We have there

    type Mutator[T] = (TreeTransform, T, Context) => TreeTransform

It turns out that then Mutator[X] for some type X did not typecheck because
the typer got confused what were the type parameters of the Function3 type on
Mutator's RHS. The fix adds a case to handle RefinedTypes that bind type
parameters.